Implemented delete logic and also printed success message

+++ src/main/webapp/students-list.jsp
@@ -34,6 +34,14 @@
 <body>
 <h2>MVC CRUD APPLICATION</h2>
 <h5>SERVLET + JSP + JDBC</h5>
+<%
+    String success =(String)request.getParameter("success");
+    if(success!=null){
+%>
+<p style='color:green;'><%=success%></p>
+<%
+    }
+%>
 
 <a href="students?action=add">Add Student</a>
 
@@ -63,7 +71,7 @@
         <td><%= s.getMobile() %></td>
         <td>
             <a href="students?action=edit&id=<%=s.getId()%>">Update</a>
-            <a href="students?action=delete&id=<%=s.getId()%>">Delete</a>
+            <a href="students?action=delete&id=<%=s.getId()%>" onclick="return confirm('Are you sure you want to delete the record')">Delete</a>
         </td>
     </tr>
     <%
